Prepare import of bundle products so it can be handled in sub-modules.
It does not implement the import of bundle products, but only add a hook ConnectorUnit to implement it.
It imports the dependencies of the bundle though.

The PHP part of the connector has been updated in order to add the bundle information, so it needs to be updated on the Magento instance.