Prevent to export the tracking number twice
When the tracking number was written after the picking has been set to
'done', but before the job to export the picking was executed, the job
that export the picking was exporting the tracking number, but another
job was created to export the tracking number as well.
Now, the picking' job only exports the tracking number if it was already
there when the picking was set to 'done' (because in that case, the
second job is not created).