[FIX] magento.website: map backend_id only on create
Without this fix when we synchronize the metadata of a backend it
triggers the recompute of 'magento.res.partner'.backend_id field
(related stored) because the 'backend_id' is updated on
'magento.website'.
So with a high number of bindings of partners in the database, this
recompute consume a huge amount of memory.

Also, a 'magento.website' should not change its backend with time, so
better to enforce this rule.